+++ src/shaders/mesh.glsl
@@ -43,8 +43,7 @@ struct InstanceData {
 };
 
 struct Vert {
-	vec4 position;
-	
+	f16vec4 position;
 	f16vec2 texcoord;
 	f16vec2 normal;
 	f16vec4 tangent;
